+++ src/main/java/structures/ValidationTuple.java
@@ -2,4 +2,4 @@ package structures;
 
 import java.util.HashMap;
 
-public record ValidationTuple(HashMap<Integer, Long> attributeGroup, Long hash) {}
+public record ValidationTuple(HashMap<Integer, Long> attributeGroup, int hash) {}
